+++ database/migrations/2022_12_20_193306_create_categories_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
 		Schema::create('categories', function (Blueprint $table) {
 			$table->id();
 			$table->uuid('uuid');
-			$table->unsignedBigInteger('tenant_id');
+			$table->unsignedInteger('tenant_id');
 			$table->string('name');
 			$table->string('url');
 			$table->text('description')->nullable();
