+++ database/migrations/2022_12_22_161705_create_products_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
 		Schema::create('products', function (Blueprint $table) {
 			$table->id();
 			$table->uuid('uuid');
-			$table->unsignedBigInteger('tenant_id');
+			$table->unsignedInteger('tenant_id');
 			$table->string('title');
 			$table->string('flag');
 			$table->string('image')->nullable();
